ANN with Backpropagation (in Python)
The content of this file is used for learning purposes only. It visualizes and implements most of the math behind a simple Artificial Neural Network + Backpropagation by just using Python (+ Numpy) instead of common libraries or frameworks, which have these functions already integrated. 

Many thanks to Patrik Szepesi, who explains all the real math behind it in great detail on the platform ZTM.